--- pages/payments/pageSettings.ts
@@ -1,11 +1,0 @@
-export type PaymentDataInputValues = {
-  cardNumber: number;
-  securityCode: number;
-  familyName: string;
-  givenName: string;
-  postCode: string;
-  address: string;
-  buildingName: string;
-  dateOfExpiry__year: number;
-  dateOfExpiry__month: number;
-};
